Task 1 - Some Temporary Animation Work Added
