+++ src/app/animation/animation.component.html
@@ -1,0 +1,68 @@
+<ng-container class="content1" *ngIf="showContent1">
+  <div class="logo">
+    <div class="row">
+      <div class="col-5"></div>
+      <div class="col-4 logoCol2">
+        <img src="../../assets/images/logo.jpeg" class="logoImage" />
+      </div>
+      <div class="col-3"></div>
+    </div>
+  </div>
+</ng-container>
+
+<div class="content3" *ngIf="showContent3">
+  <app-home-page></app-home-page>
+</div>
+
+<div class="content2" *ngIf="showContent2 || showContent3">
+  <div class="logo" id="animateContent2">
+    <div class="row">
+      <div class="col-5"></div>
+      <div class="col-4 logoCol2">
+        <img
+          src="../../assets/images/logo.jpeg"
+          class="myLogoImage"
+          id="myImage"
+        />
+      </div>
+      <div class="col-3"></div>
+    </div>
+  </div>
+  <div class="row row1" *ngIf="showContent2"></div>
+  <div class="row row2" *ngIf="showContent2">
+    <div class="col-4 col1"></div>
+    <div class="col-6">
+      <canvas #canvas id="canvas"></canvas>
+    </div>
+    <div class="col-2"></div>
+  </div>
+  <div class="row row3" *ngIf="showContent2"></div>
+
+  <div class="wrapper row" *ngIf="showContent2">
+    <div class="col-3"></div>
+    <div class="col-9">
+      <div class="slogan">
+        Exploring The World in of START YOUR JOURNEY WITH AAMEEQ Virtual Reality
+      </div>
+    </div>
+  </div>
+</div>
+
+  <!-- <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 1000 100">
+    <circle
+      cx="22.82"
+      cy="37.27"
+      r="35"
+      stroke="yellow"
+      stroke-miterlimit="10"
+    />
+    <path fill="none" d="M49.04 39.59h49.83-49.83z" />
+    <text
+      transform="translate(27 40.261)"
+      fill="black"
+      font-size="25"
+      font-family="MyriadPro-Regular"
+    >
+      Exploring The World in of START YOUR JOURNEY WITH AAMEEQ Virtual Reality
+    </text>
+  </svg> -->
