+++ src/app/app-routing.module.ts
@@ -1,7 +1,18 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
+import { AnimationComponent } from './animation/animation.component';
+import { HomePageComponent } from './home-page/home-page.component';
 
-const routes: Routes = [];
+const routes: Routes = [
+  {
+    path: "",
+    component: AnimationComponent
+  },
+  {
+    path: "home",
+    component: HomePageComponent
+  }
+];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
